requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
django-crontab==0.7.1
django-dbbackup==4.1.0
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
validate-email-address==1
